:construction: modification for Mediator Design Pattern and code refactoring
