+++ src/main/java/main/java/edu/miu/cs/cs525/reversi/common/BoardInfo.java
@@ -3,7 +3,6 @@ package main.java.edu.miu.cs.cs525.reversi.common;
 import main.java.edu.miu.cs.cs525.reversi.ReversiSingleton;
 import main.java.edu.miu.cs.cs525.reversi.mediator.BoardEnum;
 import main.java.edu.miu.cs.cs525.reversi.mediator.MediatorService;
-import main.java.edu.miu.cs.cs525.reversi.monitor.ShowCurrentPlayer;
 
 public class BoardInfo implements MediatorService {
 
@@ -42,19 +41,7 @@ public class BoardInfo implements MediatorService {
 		moveCount = src.moveCount;
 		validMoveCount = src.validMoveCount;
 	}
-	@Override
-	public void changePlayer(ShowCurrentPlayer showCurrentPlayer) {
-		ReversiSingleton.setCurrentPlayer(showCurrentPlayer);
-	}
-	@Override
-	public void copyBoard(int[][] src, int[][] dest) {
-		int i, j;
-		for (i = 0; i < BoardEnum.ROW_COUNT.value(); i++) {
-			for (j = 0; j < BoardEnum.COL_COUNT.value(); j++) {
-				dest[i][j] = src[i][j];
-			}
-		}
-	}
+
 	@Override
 	public void makeEmpty() {
 		int i, j;
@@ -67,6 +54,7 @@ public class BoardInfo implements MediatorService {
 		validMoveCount = 0;
 		turn = BoardEnum.NO_GAME.value();
 	}
+
 	@Override
 	public void initBoard() {
 		makeEmpty();
@@ -104,6 +92,7 @@ public class BoardInfo implements MediatorService {
 		}
 	}
 
+	@Override
 	public void changeTurn() {
 		if (turn == BoardEnum.PLAYER_BLACK.value()) {
 			turn = BoardEnum.PLAYER_WHITE.value();
@@ -112,6 +101,7 @@ public class BoardInfo implements MediatorService {
 		}
 	}
 
+	@Override
 	public boolean isValidMove(int r, int c) {
 		int iD, jD, dCount = 0;
 		if (turn == BoardEnum.NO_GAME.value() || turn == BoardEnum.GAME_OVER.value()
@@ -137,6 +127,7 @@ public class BoardInfo implements MediatorService {
 		}
 	}
 
+	@Override
 	public AnimationMatrix calculateMoveAnimation(int r, int c) {
 		AnimationMatrix anim = new AnimationMatrix();
 		int iD, jD, dCount = 0;
@@ -171,7 +162,7 @@ public class BoardInfo implements MediatorService {
 		return anim;
 	}
 
-	public int performMove(int r, int c) {
+	private int performMove(int r, int c) {
 		// Performs a move with a trick :
 		// calculates move animation & then performs animation at once !
 		AnimationMatrix am = calculateMoveAnimation(r, c);
@@ -179,10 +170,12 @@ public class BoardInfo implements MediatorService {
 		return correctTurn();
 	}
 
+	@Override
 	public int performMove(Location l) {
 		return performMove(l.row, l.column);
 	}
 
+	@Override
 	public String getTurnString() {
 		if (turn == BoardEnum.PLAYER_BLACK.value()) {
 			changePlayer(ReversiSingleton.getBlackPlayer());
@@ -202,6 +195,7 @@ public class BoardInfo implements MediatorService {
 		}
 	}
 
+	@Override
 	public int getPieceCount(int player) {
 		int c = 0;
 		int i, j;
@@ -230,6 +224,7 @@ public class BoardInfo implements MediatorService {
 		}
 	}
 
+	@Override
 	public BoardMatrix getGainMatrix() {
 		BoardMatrix m = new BoardMatrix();
 		int r, c, iD, jD;
@@ -260,6 +255,7 @@ public class BoardInfo implements MediatorService {
 		return m;
 	}
 
+	@Override
 	public BoardMatrix getPossibleMovesMatrix() {
 		BoardMatrix m = getGainMatrix();
 		int i, j;
@@ -275,6 +271,7 @@ public class BoardInfo implements MediatorService {
 		return m;
 	}
 
+	@Override
 	public int correctTurn() {
 		BoardMatrix m = getGainMatrix();
 		if (m.getMax() > 0) {
@@ -387,25 +384,17 @@ public class BoardInfo implements MediatorService {
 		}
 	}
 
-	public String getPossibleMoves() {
-		String s = "";
-		BoardMatrix m = getGainMatrix();
-		int i, j;
-		for (i = 0; i < BoardEnum.ROW_COUNT.value(); i++) {
-			for (j = 0; j < BoardEnum.COL_COUNT.value(); j++) {
-				if (m.get(i, j) > 0) {
-					Location l = new Location(i, j);
-					s = s + l.getStandardForm() + ",";
-				}
-			}
+	private boolean isEmpty(int i, int j) {
+		if (i < 0 || i >= BoardEnum.ROW_COUNT.value() || j < 0 || j >= BoardEnum.COL_COUNT.value()) {
+			return false;
+		} else if (board[i][j] == BoardEnum.EMPTY.value()) {
+			return true;
+		} else {
+			return false;
 		}
-		if (s.length() > 0) {
-			s = s.substring(0, s.length() - 1);
-		}
-		return s;
 	}
 
-	public int countCornerStablePieces(int player, Location corner, int rMAx, int cMax) {
+	private int countCornerStablePieces(int player, Location corner, int rMAx, int cMax) {
 		int i, j, c = 0, rm = 0, cm = 0, rec = 0;
 		boolean e;
 		Location dir = new Location((corner.row <= BoardEnum.ROW_COUNT.value() / 2 ? +1 : -1),
@@ -458,6 +447,7 @@ public class BoardInfo implements MediatorService {
 		return c;
 	}
 
+	
 	private int countRepetitions(int player, Location corner1, Location corner2, int n) {
 		int c = 0;
 		Location dir = new Location();
@@ -491,6 +481,7 @@ public class BoardInfo implements MediatorService {
 		return c;
 	}
 
+	@Override
 	public int countStablePieces(int player) {
 		int sum = 0;
 		sum += countCornerStablePieces(player, new Location(0, 0), BoardEnum.ROW_COUNT.value(),
@@ -505,26 +496,9 @@ public class BoardInfo implements MediatorService {
 		return sum;
 	}
 
-	public int getOpponent(int player) {
-		if (player == BoardEnum.PLAYER_BLACK.value()) {
-			return BoardEnum.PLAYER_WHITE.value();
-		} else {
-			return BoardEnum.PLAYER_BLACK.value();
-		}
-	}
-
-	private boolean isEmpty(int i, int j) {
-		if (i < 0 || i >= BoardEnum.ROW_COUNT.value() || j < 0 || j >= BoardEnum.COL_COUNT.value()) {
-			return false;
-		} else if (board[i][j] == BoardEnum.EMPTY.value()) {
-			return true;
-		} else {
-			return false;
-		}
-	}
-
+	@Override
 	public int countFrontier(int player) {
-	  	int f = 0;
+		int f = 0;
 		int i, j;
 		for (i = 0; i < BoardEnum.ROW_COUNT.value(); i++) {
 			for (j = 0; j < BoardEnum.COL_COUNT.value(); j++) {
